refactor(e2e): extract ceColors sampling helper shared by three dark-mode tests
